sync service, refactoring
