+++ Code/VFSPrototype/VFSBase/Callbacks/CopyCallbacks.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace VFSBase.Interfaces
+namespace VFSBase.Callbacks
 {
     public class CopyCallbacks : CallbacksBase
     {
