+++ Code/VFSPrototype/VFSBase/Callbacks/SynchronizationCallbacks.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using VFSBase.DiskServiceReference;
+
+namespace VFSBase.Synchronization
+{
+    public class SynchronizationCallbacks
+    {
+        public SynchronizationCallbacks(Action<SynchronizationState> stateChanged)
+        {
+            StateChanged = stateChanged ?? (s => { });
+        }
+
+        public Action<SynchronizationState> StateChanged { get; private set; }
+    }
+}
