+++ Code/VFSPrototype/VFSBase/Interfaces/FileSystemChangedEventArgs.cs
@@ -4,11 +4,5 @@ namespace VFSBase.Interfaces
 {
     internal class FileSystemChangedEventArgs : EventArgs
     {
-        public FileSystemChangedEventArgs(string location)
-        {
-            Location = location;
-        }
-
-        public string Location { get; private set; }
     }
 }
