+++ Code/VFSPrototype/VFSBase/Interfaces/ISynchronizationService.cs
@@ -1,6 +1,7 @@
-﻿namespace VFSBase.Implementation
+﻿namespace VFSBase.Interfaces
 {
     public interface ISynchronizationService
     {
+        void Synchronize();
     }
 }
